backend/requirements.txt
# Dark Knight Phantom SIEM - Backend Requirements
# Django Core
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
django-filter>=23.5

# API Documentation
drf-spectacular>=0.27

# Date/Time handling
python-dateutil>=2.8

# Async & Real-time
channels>=4.0
channels-redis>=4.1

# Task Queue
celery>=5.3
redis>=5.0

# Database (Production)
psycopg2-binary>=2.9

# Security
PyJWT>=2.8
bcrypt>=4.1

# Utilities
python-json-logger>=2.0
pyyaml>=6.0



